+++ zephyr/modules/owntech_communication/zephyr/public_api/CanCommunication.h
@@ -46,7 +46,7 @@ public:
 
 	static void setCanNodeAddr(uint16_t addr);
 	static void setCtrlEnable(bool enable);
-	static void setCtrlReference(float32_t reference);
+	static void setCtrlReference(uint16_t reference);
 	static void setBroadcastPeriod(uint16_t time_100_ms);
 	static void setControlPeriod(uint16_t time_100_ms);
